requirements.txt
asgiref==3.2.10
boto==2.49.0
boto3==1.14.34
botocore==1.17.34
Django==2.2.14
django-debug-toolbar==2.2
django-filebrowser==3.13.1
django-filter==2.3.0
django-grappelli==2.14.2
django-ses==1.0.2
django-tinymce==2.9.0
djangorestframework==3.11.0
docutils==0.15.2
future==0.18.2
importlib-metadata==1.7.0
jmespath==0.10.0
Markdown==3.2.2
mysqlclient==2.0.1
olefile==0.46
Pillow==8.4.0
python-dateutil==2.8.1
pytz==2020.1
s3transfer==0.3.3
six==1.15.0
sqlparse==0.3.1
urllib3==1.25.10
zipp==1.2.0
